+++ com.zsmartsystems.zigbee/src/main/java/com/zsmartsystems/zigbee/zcl/clusters/levelcontrol/ZclLevelControlExtensionField.java
@@ -80,13 +80,13 @@ public class ZclLevelControlExtensionField extends ExtensionFieldSet {
 
     @Override
     public void deserialize(final ZigBeeDeserializer deserializer) {
-        clusterId = (int) deserializer.readZigBeeType(ZclDataType.UNSIGNED_16_BIT_INTEGER);
-        int size = (int) deserializer.readZigBeeType(ZclDataType.UNSIGNED_8_BIT_INTEGER);
+        clusterId = deserializer.readZigBeeType(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+        int size = deserializer.readZigBeeType(ZclDataType.UNSIGNED_8_BIT_INTEGER);
         if (size >= 1) {
-            currentLevel = (Integer) deserializer.readZigBeeType(ZclDataType.UNSIGNED_8_BIT_INTEGER);
+            currentLevel = deserializer.readZigBeeType(ZclDataType.UNSIGNED_8_BIT_INTEGER);
         }
         if (size >= 3) {
-            currentFrequency = (Integer) deserializer.readZigBeeType(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+            currentFrequency = deserializer.readZigBeeType(ZclDataType.UNSIGNED_16_BIT_INTEGER);
         }
     }
