+++ com.zsmartsystems.zigbee/src/main/java/com/zsmartsystems/zigbee/zcl/clusters/smartenergytunneling/TunnelClosureNotification.java
@@ -117,7 +117,7 @@ public class TunnelClosureNotification extends ZclSmartEnergyTunnelingCommand {
 
     @Override
     public void deserialize(final ZclFieldDeserializer deserializer) {
-        tunnelId = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+        tunnelId = deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
     }
 
     @Override
